chore: enhance typo check workflow (#667)
This PR adds a typo check to CI using the typos GitHub Action. I also
created a .typos.toml config file so we can ignore certain words that
are either project specific, technical terms, or intentionally spelled
that way. Without this config, the action would flag a lot of false
positives, so this should make the checks more useful and less noisy.

If you notice any words in the whitelist that actually are typos and
should be fixed, feel free to update the config or let me know. For
example, I wasn’t sure whether to change “qualifed” to “qualified”,
since it appears in a lot of places and might be intentional.

https://github.com/crate-ci/typos

Signed-off-by: atilsensalduz <atil.sensalduz@gmail.com>
Co-authored-by: Romain Marcadier <romain@datadoghq.com>