fix: flags only partially processed
When using `go test` and providing the test package set before some flag, the
subsequent flags were not correctly processed, resulting in particular in the
`-coverpkg` flag not being appropriately expanded when necessary. This resulted
in relative patterns being passed to child builds which is the cause of some
build ID mismatch errors.

This was caused by an incorrect assumption that all `go` building commands
demand flags are presented before any positional argument, which is at least not
true for `go test` (and likely also for `go run`).